requirements.txt
smolagents==1.13.0
gradio>=4.0.0,<5.0.0
smolagents[litellm]
pandas
python-dotenv
pyyaml
requests
beautifulsoup4
tqdm>=4.65.0
huggingface-hub>=0.20.0
